[lib] Separate procedures and triggers from system/database
